feat: Rows 제거
- Ladder가 List<Row>에 대한 일급컬렉션 그 자체의 역할을 하고 있다고 판단
