+++ src/test/java/nextstep/ladder/domain/ladder/LadderTest.java
@@ -9,19 +9,19 @@ import java.util.Collections;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class RowsTest {
+public class LadderTest {
 
-    @DisplayName("Rows는 높이 만큼의 행을 가진다.")
+    @DisplayName("사다리는 높이 만큼의 행을 가진다.")
     @Test
     void constructor() {
-        assertThat(new Rows(new Height(5), new Count(3), () -> true).height())
+        assertThat(new Ladder(new Height(5), new Count(3), () -> true).height())
                 .isEqualTo(new Height(5));
     }
 
-    @DisplayName("Rows는 반드시 Row를 한 개 이상 가져야 한다.")
+    @DisplayName("사다리는 반드시 Row를 한 개 이상 가져야 한다.")
     @Test
     void empty_constructor() {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> new Rows(Collections.emptyList()));
+                .isThrownBy(() -> new Ladder(Collections.emptyList()));
     }
 }
